Fix du bug : niveau en chaine de caractères qui ne s'affiche pas pour les langues

+++ database/migrations/2023_07_07_155742_create_language_table.php
@@ -17,7 +17,7 @@ class CreateLanguageTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('string_level')->nullable();
-            $table->integer('level')->unsigned()->nullable();
+            $table->float('level')->unsigned()->nullable();
             $table->foreignId('cv_id')->constrained()->onDelete('cascade');
         });
     }
